refactor: integrate RichTextEditorSkeleton into RichTextEditor and remove separate file
